# apps/api/security/openapi.json
{
  "components": {
    "schemas": {
...
      "DockerContainerSchema": {
        "properties": {
          "id": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "state": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "names",
          "image",
          "status",
          "state"
        ],
        "type": "object"
      },
      "DockerContainersEnvelope": {
        "properties": {
          "containers": {
            "items": {
              "$ref": "#/components/schemas/DockerContainerSchema"
            },
            "type": "array"
          }
        },
        "required": [
          "containers"
        ],
        "type": "object"
      },
...
      "UserEnvelope": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          }
        },
        "required": [
          "user"
        ],
        "type": "object"
      },
      "UserListEnvelope": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserSchema"
            },
            "type": "array"
          }
        },
        "required": [
          "users"
        ],
        "type": "object"
      },
      "UserSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "direct_max_instances": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "group_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_admin": {
            "type": "boolean"
          },
          "tier": {
            "format": "int32",
            "type": "integer"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at",
          "group_ids",
          "is_admin",
          "tier"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI spec for the 17 security-fuzzable endpoints (see .scratch/security-fuzzing). The 8 admin-gated operations are tagged `admin-gated` for the low-privilege RBAC-boundary pass. The spec is export-only and never served.",
    "license": {
      "name": ""
    },
    "title": "OpenWorkspace API — security fuzz surface",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/docker/containers": {
      "get": {
        "operationId": "list_docker_containers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DockerContainersEnvelope"
                }
              }
            },
            "description": "docker containers visible to the host"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_docker or admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "list_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListEnvelope"
                }
              }
            },
            "description": "user catalog with policy overrides"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "requires can_manage_users or admin"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "get_user",
        "parameters": [
          {
            "description": "user uuid",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelope"
                }
              }
            },
            "description": "user detail with policy overrides"
          },
          "400": {
            "description": "invalid uuid"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "403": {
            "description": "not allowed to view this user"
          },
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "admin-gated"
        ]
      }
    },
    "/api/vnc/verify": {
      "get": {
        "operationId": "vnc_verify",
        "responses": {
          "200": {
            "description": "verified; VNC websockify forwarding headers returned"
          },
          "401": {
            "description": "missing or invalid ow_token / forwarded uri"
          },
          "404": {
            "description": "unknown or non-running instance"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "vnc"
        ]
      }
    },
...
  }
}